Рефактор теста по успешной регистрации пользователя
